Reports find Sales,Returns,InvoiceID in daterange
Added findIDsBetweenDates function to ReportsImpl class.
This searches a table for ID's whose dates fall in a specified period.
Returns a vector of all the ID's.
Works with Sales, Returns, and ID's, just change the Table type used in
the paramaters when using the functions.

Added four testcases for the function